refactored FCM service class, moved utility functions to separated classes
